[1/2] SystemUI: Add time-context headers to the notification header
Also squashed the following commits:
SystemUI: Fix notification header disappearing on tablets
https://github.com/omnirom/android_frameworks_base/commit/2f2ad3b28cb388b140db21233265614841e6ac73

SystemUI: Cleanup and enhance notification drawer header
https://github.com/omnirom/android_frameworks_base/commit/7a2804872ae5364de1d0431952ac4fd7587ac3e1

This commit adds image headers to the notification panel header,
based on the date/time. There are pictures for morning, afternoon
and night, as well as special events like Christmas and new
year's eve.
More will be coming, with eventually a way to use an external
(as in, Internet) provider for those pictures.

Change-Id: Ie3321e9a3e0637d6ba8094dd46aee36c41354713

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java