[1/2] Base: On-The-Go mode
Also squashed:

Base: On-The-Go: Allow to toggle between front and back camera
https://github.com/NamelessRom/android_frameworks_base/commit/e28324c182f778cae300340b0c8760076961fa06

Base: On-The-Go: Several fixes
https://github.com/NamelessRom/android_frameworks_base/commit/755e04c2ca9251e9ab87da2319da34e4c97b6e3a

Base: On-The-Go: Overlay status- and navigation bar as well
https://github.com/NamelessRom/android_frameworks_base/commit/ede7578328baeeedcb1873b2c8061b0e21cf8e09

Base: On-The-Go: Change the way of changing the camera mode
https://github.com/NamelessRom/android_frameworks_base/commit/73509791602ccca66b75a20ec3a36b7d3aae7bf8

Base: On-The-Go: Allow to restart service on the fly
https://github.com/NamelessRom/android_frameworks_base/commit/19c86316bb34edb71eb1bb39ecbbd6cce6d85ca4

[1/2] Base: On-The-Go: Add quicktile
https://github.com/NamelessRom/android_frameworks_base/commit/5c63f5a0c06b6c201bd732cc72eea0bd2ba773d9

Base: On-The-Go: Switch Licenses to GPLv3
https://github.com/NamelessRom/android_frameworks_base/commit/2a15bcec684ec501d16daf6bacaa3892865f7b0a

[1/2] Base: On-The-Go: Hide on devices without camera
https://github.com/NamelessRom/android_frameworks_base/commit/1fa20bc7d32a83a25ae69641b4815abe80a402cf

Base: On-The-Go: Turn stop and start service if screen turns off and on
https://github.com/NamelessRom/android_frameworks_base/commit/90ac0f63af927837a79937b0d3c4061ac6fc9f17

On-The-Go turns on the camera of your phone and streams it to an overlay,
which is drawn over everything.
this makes you feel like your display is transparent, as you see the content
behind it.

its really useful when you are using your phone while walking, thats why
we called it "On-The-Go"

allows to change transparency of the overlay via the notification.

have fun!

Change-Id: I3efccb5fac41a946f6cb4992e9181422980ee5e4

Conflicts:
	core/java/android/provider/Settings.java
	core/res/res/values/custom_strings.xml
	core/res/res/values/custom_symbols.xml
	packages/SystemUI/AndroidManifest.xml
	packages/SystemUI/res/values/custom_strings.xml
	policy/src/com/android/internal/policy/impl/GlobalActions.java