Add customization for Quick Settings [1/2]
Squashed:
QS: Don't show inversion and hotspot tiles by default
https://github.com/CyanogenMod/android_frameworks_base/commit/c770874228a05abfa99e31514536bc6976caa174

Allow Quick Settings tiles to be arranged any way you want, delete them or add some.

I removed the "dual" status that was hardcoded into Wifi and BT tiles and added it
as a variable which is true when the tile is on the first row.

Also, the first row is now hardcoded to have only 2 items to follow the same design as before.

Inversion and Hotspot tile are no longer contextual and can be added, rmeoved or rearranged at will.

Settings: http://review.cyanogenmod.org/83418

Change-Id: Ie596a54a5c4eb777782fb047496cfcd0a6cd2e89

Conflicts:
	core/java/android/provider/Settings.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/QSTileHost.java

Conflicts:
	core/java/android/provider/Settings.java