Themes: Use mRealPackage, when available, for idmapping
Some packages replace existing apps and take on their package name,
i.e. com.cyanogenmod.trebuchet -> com.android.launcher3.  When this
happens pkg.packageName becomes the name of the package it is
overriding and mRealPackage will be the original package name.

Because it is possible for the resources in the new package to not
match up with the package it replaced, we need to perform the idmap
using the original package name.

Change-Id: Ic22c7005274ae5c4ea29d9a4b29f9b9f40719e68